pyproject.toml
[project]
name = "linkly"
version = "0.1.0"
description = "Linkly is a fast, lightweight URL shortening service."
readme = "README.md"
requires-python = ">=3.12"

authors = [
  { name="Dorna Raj Gyawali", email="dronarajgyawali@gmail.com" }
]

dependencies = [
    "aioredis>=2.0.1",
    "black>=25.1.0",
    "fastapi>=0.115.13",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "uvicorn>=0.34.3",
]
